% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/240Clean_add_chain_ident_remove_prefix.R
\name{add_chain_ident_remove_prefix}
\alias{add_chain_ident_remove_prefix}
\title{Title}
\usage{
add_chain_ident_remove_prefix(data)
}
\value{
Return the input data with a column added indicating the chain,
and with the prefix "TRA" and "TRB" removed.
}
\description{
Title
}
\seealso{
Other Cleaning functions: 
\code{\link{allele_names}()},
\code{\link{pivot_longer_TCR_sequences}()},
\code{\link{pivot_longer_allele}()},
\code{\link{remove_unnecessary_binders}()},
\code{\link{run_all_clean}()},
\code{\link{split_TCR_sequences}()}
}
\concept{Cleaning functions}
