% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/260Clean_allele_names.R
\name{allele_names}
\alias{allele_names}
\title{Title}
\usage{
allele_names(data)
}
\value{
Return a data frame with the "allele_info" split into three columns:
\itemize{
\item "allele"
\item "peptide"
\item "peptide_source"
While keeping the original column.
}
}
\description{
Title
}
\seealso{
Other Cleaning functions: 
\code{\link{add_chain_ident_remove_prefix}()},
\code{\link{pivot_longer_TCR_sequences}()},
\code{\link{pivot_longer_allele}()},
\code{\link{remove_unnecessary_binders}()},
\code{\link{run_all_clean}()},
\code{\link{split_TCR_sequences}()}
}
\concept{Cleaning functions}
