% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/460Model_alpha_beta_sequence_consistency.R
\name{alpha_beta_consistency}
\alias{alpha_beta_consistency}
\title{Title}
\usage{
alpha_beta_consistency(.data, identifier = barcode)
}
\arguments{
\item{.data}{An prepared data frame ready for modelling.}

\item{identifier}{A variable in the data frame with which the data set should
be distinct by to avoid a bias towards experiments with many observations.
Needs to be specified, even if set to default.}
}
\value{
Returns a bar plot with the relative distinctiveness of all alpha
and beta chains. Meaning a lower value, means less distinctive sequences
compared to the total number of that chain.
}
\description{
Title
}
\seealso{
Other Modelling functions: 
\code{\link{alpha_beta_pair_distribution}()},
\code{\link{logo_plot}()},
\code{\link{multiple_relevant_binders}()},
\code{\link{relevant_binder_frequency_plot}()},
\code{\link{summarise_with_filter}()}
}
\concept{Modelling functions}
