% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/450Model_alpha_beta_pair_distributions.R
\name{alpha_beta_pair_distribution}
\alias{alpha_beta_pair_distribution}
\title{Title}
\usage{
alpha_beta_pair_distribution(.data, pMHC_filter = FALSE)
}
\arguments{
\item{.data}{An augmented data frame ready for modelling.}

\item{pMHC_filter}{A pMHC to use for filtering, to only include that specific
pMHC from the data set. Default is \code{FALSE}, i.e. no filtering.}
}
\value{
Returns a bar plot with a frequency distribution of the different
combinations of alpha- beta pairs.
}
\description{
Title
}
\seealso{
Other Modelling functions: 
\code{\link{logo_plot}()},
\code{\link{multiple_specific_binders}()},
\code{\link{relevant_binder_frequency_plot}()},
\code{\link{summarise_with_filter}()}
}
\concept{Modelling functions}
