% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aux_alleles_status_each_donor}
\alias{aux_alleles_status_each_donor}
\title{Auxiliary data set with alleles and disease status}
\format{
A data frame with 4 rows and 7 columns
}
\source{
\url{https://www.10xgenomics.com/resources/document-library/a14cde}
}
\usage{
aux_alleles_status_each_donor
}
\description{
The data frame contains information of which allele group and HLA protein
each donor has along with information of disease status.
}
\keyword{datasets}
