% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/470Model_boxplot_protein_expressions.R
\name{boxplot_protein_expressions}
\alias{boxplot_protein_expressions}
\title{Umap of the protein expression levels}
\usage{
boxplot_protein_expressions(.data, plot_protein = "CD4")
}
\arguments{
\item{.data}{An prepared data frame ready for modelling.}
}
\value{
Boxplots for each donor and protein present in the data set
}
\description{
\code{boxplot_protein_expressions()} takes a prepared data frame, \code{.data}, and
pivots the data longer, to allow for a boxplot to show protein
expression levels.
}
\examples{
# A prepared data frame is simply piped through the function:
data_combined_tidy \%>\%
    boxplot_protein_expressions()

# The protein to be plotted can be changed by supplying it as a string:
data_combined_tidy \%>\%
    boxplot_protein_expressions(plot_protein = "CD8a")
}
\seealso{
Other Modelling functions: 
\code{\link{alpha_beta_distinctiveness}()},
\code{\link{alpha_beta_pair_distribution}()},
\code{\link{count_binding_pr_allele}()},
\code{\link{percentage_rows_kept}()},
\code{\link{relevant_binder_frequency_plot}()},
\code{\link{summarise_with_filter}()},
\code{\link{umap_of_protein_expressions}()}
}
\concept{Modelling functions}
