% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/450Model_count_binding_pr_allele.R
\name{count_binding_pr_allele}
\alias{count_binding_pr_allele}
\title{Scatter plot showing relevant binder count for each allele type}
\usage{
count_binding_pr_allele(.data, identifier = barcode)
}
\arguments{
\item{.data}{An prepared data frame ready for modelling.}

\item{identifier}{A variable in the data frame with which the data set should
be distinct by to avoid a bias towards experiments with many observations.}
}
\value{
A scatter plot showing the count of relevant binders for each allele.
}
\description{
\code{count_binding_pr_allele()} takes a prepared data frame, \code{.data}, and
count the number of relevant binders that exist for each allele present.
This is done by making pMHC unique after grouping on \code{identifier} and then
counting.
}
\examples{
# A prepared data frame is simply piped through the function:
data_donor_four_tidy \%>\%
    count_binding_pr_allele()

}
\seealso{
Other Modelling functions: 
\code{\link{HLA_match_distribution}()},
\code{\link{alpha_beta_distinctiveness}()},
\code{\link{alpha_beta_pair_distribution}()},
\code{\link{relevant_binder_frequency_plot}()},
\code{\link{summarise_with_filter}()}
}
\concept{Modelling functions}
