% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_donor_four}
\alias{data_donor_four}
\title{Prepared data from donor four}
\format{
A data frame with 190k rows and 27 columns
}
\source{
Prepared data from \code{vdj_v1_hs_aggregated_donor4_binarized_matrix}
}
\usage{
data_donor_four
}
\description{
The data frame contains a prepared version of the data from donor four
}
\keyword{datasets}
