% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_donor_one}
\alias{data_donor_one}
\title{Prepared data from donor one}
\format{
A data frame with 500k rows and 27 columns
}
\source{
Prepared data from \code{vdj_v1_hs_aggregated_donor1_binarized_matrix}
}
\usage{
data_donor_one
}
\description{
The data frame contains a prepared version of the data from donor one
}
\keyword{datasets}
