% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_donor_one_mock}
\alias{data_donor_one_mock}
\title{Mock data from donor one, top 1000 rows}
\format{
A data frame with 1k rows and 27 columns
}
\source{
Prepared data from \code{data_donor_one_tidy}
}
\usage{
data_donor_one_mock
}
\description{
The data frame contains data from donor one, but only the top 1000 rows.
The purpose is to use this mock data in techinal report and vignette to speed
up computations.
}
\keyword{datasets}
