% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_donor_one_raw_mock}
\alias{data_donor_one_raw_mock}
\title{Mock raw data from donor one, top 1000 rows}
\format{
A data frame with 1k rows and 118 columns
}
\source{
Data from \code{data_donor_one_raw}
}
\usage{
data_donor_one_raw_mock
}
\description{
The data frame contains raw data from donor one, but only the top 1000 rows.
The purpose is to use this raw mock data in technical report and vignette to
speed up computations.
}
\keyword{datasets}
