% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_donor_three_tidy}
\alias{data_donor_three_tidy}
\title{Prepared data from donor three}
\format{
A data frame with 580k rows and 27 columns
}
\source{
Prepared data from \code{data_donor_three_raw}
}
\usage{
data_donor_three_tidy
}
\description{
The data frame contains a prepared version of the data from donor three
}
\keyword{datasets}
