% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_donor_two}
\alias{data_donor_two}
\title{Prepared data from donor two}
\format{
A data frame with 800k rows and 27 columns
}
\source{
Prepared data from \code{vdj_v1_hs_aggregated_donor2_binarized_matrix}
}
\usage{
data_donor_two
}
\description{
The data frame contains a prepared version of the data from donor two
}
\keyword{datasets}
