% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_donor_two_raw}
\alias{data_donor_two_raw}
\title{CD8+ T-cells of Healthy donor 2}
\format{
A data frame with 77854 rows and 118 variables:
\describe{
\item{barcode}{basepair sequence of peptide}
\item{donor}{name of donor}
\item{cell_clono_cdr3_aa}{amino acid sequence of CDR3 alpha and beta chains}
\item{cell_clono_cdr3_nt}{nucleotide sequence of CDR3 alpha and beta chains}
}
}
\source{
\url{https://www.10xgenomics.com/resources/datasets/cd-8-plus-t-cells-of-healthy-donor-2-1-standard-3-0-2}
}
\usage{
data_donor_two_raw
}
\description{
A dataset containing roughly 80 000 TCR alpha- and beta sequences for
Donor 2 along with different coding regions.
}
\keyword{datasets}
