% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/210Clean_ensure_one_alpha_beta_pair.R
\name{ensure_one_alpha_beta_pair}
\alias{ensure_one_alpha_beta_pair}
\title{Title}
\usage{
ensure_one_alpha_beta_pair(data)
}
\arguments{
\item{data}{is a data structure which need cleaning}
}
\value{
The input data set, but only observations with exactly one TRA and TRB
}
\description{
Title
}
\seealso{
Other Cleaning functions: 
\code{\link{RunAllClean}()}
}
\concept{Cleaning functions}
