% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/210Clean_EnsureOneAlphaBetaPair.R
\name{EnsureOneAlphaBetaPair}
\alias{EnsureOneAlphaBetaPair}
\title{Title}
\usage{
EnsureOneAlphaBetaPair(data)
}
\arguments{
\item{data}{is a data structure which need cleaning}
}
\value{
The input data set, but only observations whith exactly one TRA and TRB
}
\description{
Title
}
\seealso{
Other Cleaning functions: 
\code{\link{RunAllClean}()}
}
\concept{Cleaning functions}
