% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/310Augment_evaluate_binder.R
\name{evaluate_binder}
\alias{evaluate_binder}
\title{Title}
\usage{
evaluate_binder(
  .data_clean,
  identifier = barcode,
  threshold = 10,
  non_specific_threshold = 5
)
}
\arguments{
\item{.data_clean}{}

\item{identifier}{}

\item{threshold}{}

\item{non_specific_threshold}{}
}
\description{
Title
}
