% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/480Model_HLA_match_distribution.R
\name{HLA_match_distribution}
\alias{HLA_match_distribution}
\title{Predict allele typing if missing}
\usage{
HLA_match_distribution(.data, identifier = barcode)
}
\arguments{
\item{.data}{An prepared data frame ready for modelling.}

\item{identifier}{A variable in the data frame with which the data set should
be distinct by to avoid a bias towards experiments with many observations.}
}
\value{
A bar plot showing the distribution between \code{TRUE} and \code{FALSE}.
}
\description{
\code{HLA_match_distribution()} takes a prepared data frame, \code{.data}, and plot the
distribution of whether or not the HLA-typing is a match. It is a match
if \code{allele} for a relevant binder matches the noted HLA from the article.
}
\examples{
# A prepared data frame is simply piped through the function:
data_donor_four_tidy \%>\%
    HLA_match_distribution()

}
\seealso{
Other Modelling functions: 
\code{\link{alpha_beta_consistency}()},
\code{\link{alpha_beta_pair_distribution}()},
\code{\link{count_binding_pr_allele}()},
\code{\link{logo_plot}()},
\code{\link{multiple_relevant_binders}()},
\code{\link{predict_virus_status}()},
\code{\link{relevant_binder_frequency_plot}()},
\code{\link{summarise_with_filter}()}
}
\concept{Modelling functions}
