% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/440Model_multiple_relevant_binders.R
\name{multiple_relevant_binders}
\alias{multiple_relevant_binders}
\title{Title}
\usage{
multiple_relevant_binders(.data, identifier = barcode)
}
\arguments{
\item{.data}{An augmented data frame ready for modelling.}

\item{identifier}{A variable in the data frame with which the data set should
be distinct by to avoid a bias towards experiments with many observations.
Needs to be specified, even if set to default.}
}
\value{
Returns a table with all \code{identifier}s which has at least two \code{pMHC}
marked as relevant.
}
\description{
Title
}
\seealso{
Other Modelling functions: 
\code{\link{alpha_beta_consistency}()},
\code{\link{alpha_beta_pair_distribution}()},
\code{\link{logo_plot}()},
\code{\link{relevant_binder_frequency_plot}()},
\code{\link{summarise_with_filter}()}
}
\concept{Modelling functions}
