% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/440Model_multiple_specific_binders.R
\name{multiple_specific_binders}
\alias{multiple_specific_binders}
\title{Title}
\usage{
multiple_specific_binders(.data)
}
\arguments{
\item{.data}{An augmented data frame ready for modelling.}
}
\value{
Returns a table with all \code{barcode}s which has atleast two \code{pMHC}
marked as significant.
}
\description{
Title
}
