% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{myFunction}
\alias{myFunction}
\title{A test function to setup workspace}
\usage{
myFunction(foo, bar)
}
\arguments{
\item{foo}{is a float}

\item{bar}{is also a float}
}
\value{
A float calculated from the mathematical expression in the function
}
\description{
A test function to setup workspace
}
\examples{
myFunction(1,2)
myFunction(4,9)
}
