% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/480Model_percentage_rows_kept.R
\name{percentage_rows_kept}
\alias{percentage_rows_kept}
\title{Table with percentage of rows left after filtering}
\usage{
percentage_rows_kept(.data, identifier = barcode)
}
\arguments{
\item{.data}{An prepared data frame ready for modelling.}

\item{identifier}{A variable in the data frame with which the data set should
be distinct by to avoid a bias towards experiments with many observations.}
}
\value{
A table with a percentage representing number of rows left compared
to unfiltered tidy data.
}
\description{
\code{percentage_rows_kept()} takes a prepared data frame as input, and compares
the number of rows to the unfiltered tidy data
\code{TCRSequenceFunctions::data_combined_tidy} for each donor. The comparison
is shown in a table.
}
\examples{
# A prepared data frame is simply piped through the function:
data_combined_tidy \%>\%
    percentage_rows_kept()

}
\seealso{
Other Modelling functions: 
\code{\link{alpha_beta_distinctiveness}()},
\code{\link{alpha_beta_pair_distribution}()},
\code{\link{boxplot_protein_expressions}()},
\code{\link{count_binding_pr_allele}()},
\code{\link{relevant_binder_frequency_plot}()},
\code{\link{summarise_with_filter}()},
\code{\link{umap_of_protein_expressions}()}
}
\concept{Modelling functions}
