% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250Clean_pivot_longer_allele.R
\name{pivot_longer_allele}
\alias{pivot_longer_allele}
\title{Title}
\usage{
pivot_longer_allele(data)
}
\value{
Return a pivot longer version of the data with UMI counts as rows
instead of columns.
}
\description{
Title
}
\seealso{
Other Cleaning functions: 
\code{\link{add_chain_ident_remove_prefix}()},
\code{\link{allele_names}()},
\code{\link{pivot_longer_TCR_sequences}()},
\code{\link{remove_unnecessary_binders}()},
\code{\link{run_all_clean}()},
\code{\link{split_TCR_sequences}()}
}
\concept{Cleaning functions}
