% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/480Model_predict_allele_typing.R
\name{predict_allele_typing}
\alias{predict_allele_typing}
\title{Predict allele typing if missing}
\usage{
predict_allele_typing(.data, identifier = barcode)
}
\arguments{
\item{.data}{An prepared data frame ready for modelling.}

\item{identifier}{A variable in the data frame with which the data set should
be distinct by to avoid a bias towards experiments with many observations.}
}
\value{
A string with the predicted missing allele typing.
}
\description{
\code{predict_allele_typing()} takes a prepared data frame, \code{.data}, and try to
predict the allele typing if it is missing. This is done by counting the
most prevalent allele of the relevant binders and returns the top hit
which isn't already noted as a present allele type.
}
\examples{
# A prepared data frame is simply piped through the function:
data_donor_four_tidy \%>\%
    predict_allele_typing()


}
\seealso{
Other Modelling functions: 
\code{\link{alpha_beta_consistency}()},
\code{\link{alpha_beta_pair_distribution}()},
\code{\link{count_binding_pr_allele}()},
\code{\link{logo_plot}()},
\code{\link{multiple_relevant_binders}()},
\code{\link{relevant_binder_frequency_plot}()},
\code{\link{summarise_with_filter}()}
}
\concept{Modelling functions}
