% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/490Model_predict_virus_status.R
\name{predict_virus_status}
\alias{predict_virus_status}
\title{Predict CMV and EBV status if missing}
\usage{
predict_virus_status(.data, identifier = barcode)
}
\arguments{
\item{.data}{An prepared data frame ready for modelling.}

\item{identifier}{A variable in the data frame with which the data set should
be distinct by to avoid a bias towards experiments with many observations.}
}
\value{
A string with the predicted missing CMV and EBV statuses.
}
\description{
\code{predict_virus_status()} takes a prepared data frame, \code{.data}, and try to
predict the CMV and EBV status if it is missing. This is done...
}
\examples{
# A prepared data frame is simply piped through the function:
data_donor_four_tidy \%>\%
    predict_virus_status()

}
\seealso{
Other Modelling functions: 
\code{\link{HLA_match_distribution}()},
\code{\link{alpha_beta_consistency}()},
\code{\link{alpha_beta_pair_distribution}()},
\code{\link{count_binding_pr_allele}()},
\code{\link{logo_plot}()},
\code{\link{multiple_relevant_binders}()},
\code{\link{relevant_binder_frequency_plot}()},
\code{\link{summarise_with_filter}()}
}
\concept{Modelling functions}
