% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/470Model_relevant_binder_frequency_plot.R
\name{relevant_binder_frequency_plot}
\alias{relevant_binder_frequency_plot}
\title{Interactive plot showing relevant binders between pMHC and TCR-sequences}
\usage{
relevant_binder_frequency_plot(
  .data,
  identifier = barcode,
  max_frequency = 1,
  plotly_option = TRUE
)
}
\arguments{
\item{.data}{An prepared data frame ready for modelling.}

\item{identifier}{A variable in the data frame with which the data set should
be distinct by to avoid a bias towards experiments with many observations.}

\item{max_frequency}{A float representing the maximum allowed value of frequencies.}

\item{plotly_option}{A bool to choose between ggplot or plotly. Default is
\code{TRUE} and the output is plotly. Otherwise a regular scatterplot.}
}
\value{
An interactive plotly plot with different relevant non-promiscuous pairs
and pMHC. The size of dots depends on the frequency of that specific
interaction occuring compared to other pMHC which bind the same
non-promiscuous pair.
}
\description{
\code{relevant_binder_frequency_plot()} takes a prepared data frame, \code{.data}, and
plot pMHC vs non-promiscuous TCR-sequences. Marker size represent number
of barcodes which support that specific interaction. The color is based
on concordance which is fraction of barcodes that support a given pMHC
stratified on non-promiscuous TCR-sequences. The output is by default a
plotly plot, meaning it is interactive. The x- and y-values can be
found by hovering a dot in the plot.
}
\examples{
# A prepared data frame is simply piped through the function:
data_combined_tidy \%>\%
    relevant_binder_frequency_plot()

# The maximum value of allowed frequencies can be changed:
data_combined_tidy \%>\%
    relevant_binder_frequency_plot(max_frequency = 0.8)

}
\seealso{
Other Modelling functions: 
\code{\link{alpha_beta_distinctiveness}()},
\code{\link{alpha_beta_pair_distribution}()},
\code{\link{boxplot_protein_expressions}()},
\code{\link{count_binding_pr_allele}()},
\code{\link{percentage_rows_kept}()},
\code{\link{summarise_with_filter}()},
\code{\link{umap_of_protein_expressions}()}
}
\concept{Modelling functions}
