% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/240Clean_remove_binders.R
\name{remove_binders}
\alias{remove_binders}
\title{Title}
\usage{
remove_binders(data)
}
\arguments{
\item{data}{is a data structure which need cleaning}
}
\value{
The input data, but without all the binding assessments
}
\description{
Title
}
\seealso{
Other Cleaning functions: 
\code{\link{RunAllClean}()},
\code{\link{ensure_one_alpha_beta_pair}()},
\code{\link{pivot_longer_allele}()},
\code{\link{remove_prefix}()},
\code{\link{split_TCR_sequences}()}
}
\concept{Cleaning functions}
