% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/230Clean_remove_prefix.R
\name{remove_prefix}
\alias{remove_prefix}
\title{Title}
\usage{
remove_prefix(data)
}
\arguments{
\item{data}{is a data structure which need cleaning}
}
\value{
The input data, but with the prefix "TRA" and "TRB" removed
}
\description{
Title
}
\seealso{
Other Cleaning functions: 
\code{\link{RunAllClean}()},
\code{\link{ensure_one_alpha_beta_pair}()},
\code{\link{pivot_longer_allele}()},
\code{\link{remove_binders}()},
\code{\link{split_TCR_sequences}()}
}
\concept{Cleaning functions}
