% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/210Clean_remove_unnecessary_binders.R
\name{remove_unnecessary_binders}
\alias{remove_unnecessary_binders}
\title{Title}
\usage{
remove_unnecessary_binders(
  data,
  column1 = "_binder",
  column2 = "cell_clono_cdr3_nt"
)
}
\arguments{
\item{data}{is a data structure which need cleaning}
}
\value{
Returns the input data, but without all the binding assessments and
nucleotide sequences
}
\description{
Title
}
\seealso{
Other Cleaning functions: 
\code{\link{RunAllClean}()},
\code{\link{add_chain_ident_remove_prefix}()},
\code{\link{allele_names}()},
\code{\link{pivot_longer_TCR_sequences}()},
\code{\link{pivot_longer_allele}()},
\code{\link{split_TCR_sequences}()}
}
\concept{Cleaning functions}
