% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/210Clean_remove_unnecessary_columns.R
\name{remove_unnecessary_columns}
\alias{remove_unnecessary_columns}
\title{Title}
\usage{
remove_unnecessary_columns(
  data,
  column1 = "_binder",
  column2 = "cell_clono_cdr3_nt"
)
}
\value{
Returns the input data, but without all the binding assessments and
nucleotide sequences
}
\description{
Title
}
\seealso{
Other Cleaning functions: 
\code{\link{add_chain_ident_remove_prefix}()},
\code{\link{pivot_longer_TCR_sequences}()},
\code{\link{pivot_longer_allele}()},
\code{\link{run_all_clean}()},
<<<<<<< HEAD:man/remove_unnecessary_binders.Rd
\code{\link{split_TCR_sequences}()}
=======
\code{\link{split_TCR_sequences}()},
\code{\link{tidy_allele_names}()}
>>>>>>> 8e8c7423f3b50a390b49a73d2d56a44ddf2fa606:man/remove_unnecessary_columns.Rd
}
\concept{Cleaning functions}
