% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/200Clean_run_all_clean.R
\name{run_all_clean}
\alias{run_all_clean}
\title{Run all other clean-functions}
\usage{
run_all_clean(.data)
}
\arguments{
\item{data}{is a data structure which need cleaning}
}
\value{
Return a cleaned data structure, of the same type as the input.
The output has the following differences:
\itemize{
\item Observations are removed if they do not have a TRA:TRB pair
\item Two new columns are added, one for TRA and one for TRB
\item The binding assessments are removed
}
}
\description{
\code{RunAllClean()} cleans the data structure it is given by removing
incomplete/wrong TRA:TRB sequence pairs and removing all the binding assessments.
The following functions are run:
\itemize{
\item \code{remove_unnecessary_columns}
\item \code{split_TCR_sequences}
\item \code{pivot_longer_TCR_sequences}
\item \code{add_chain_ident_remove_prefix}
\item \code{pivot_longer_allele}
\item \code{add_max_non_specific_binder}
\item \code{tidy_allele_names}
}
}
\examples{
RunAllClean(vdj_v1_hs_aggregated_donor1_binarized_matrix.rda)
}
\seealso{
Other Cleaning functions: 
\code{\link{add_chain_ident_remove_prefix}()},
\code{\link{pivot_longer_TCR_sequences}()},
\code{\link{pivot_longer_allele}()},
\code{\link{remove_unnecessary_columns}()},
\code{\link{split_TCR_sequences}()},
\code{\link{tidy_allele_names}()}
}
\concept{Cleaning functions}
