% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/200Clean_run_all_prep.R
\name{run_all_prep}
\alias{run_all_prep}
\title{Run all preperation functions}
\usage{
run_all_prep(.data)
}
\arguments{
\item{data}{is a data structure which need preparation before modelling.}
}
\value{
Return a prepared data structure, of the same type as the input.
The output has the following differences:
\itemize{
\item Observations are removed if they do not have a TRA:TRB pair
\item Two new columns are added, one for TRA and one for TRB
\item The binding assessments are removed
}
}
\description{
\code{run_all_prep()} prepares the data structure it is given by removing
incomplete/wrong TRA:TRB sequence pairs and removing all the binding assessments.
The following functions are run:
\itemize{
\item \code{remove_unnecessary_columns}
\item \code{split_TCR_sequences}
\item \code{pivot_longer_TCR_sequences}
\item \code{add_chain_ident_remove_prefix}
\item \code{pivot_longer_allele}
\item \code{add_max_non_specific_binder}
\item \code{tidy_allele_names}
}
}
\examples{
run_all_prep(vdj_v1_hs_aggregated_donor1_binarized_matrix.rda)
}
\concept{Cleaning functions}
