% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/430Model_significant_binder_frequency.R
\name{significant_binder_frequency}
\alias{significant_binder_frequency}
\title{Title}
\usage{
significant_binder_frequency(.data_aug, TCR_sequence_input = "CAARVRGFGNVLHC")
}
\description{
Title
}
