% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/410Model_summarise_binders.R
\name{summarise_binders}
\alias{summarise_binders}
\title{Title}
\usage{
summarise_binders(.data, ..., distinct_by = barcode)
}
\arguments{
\item{distinct_by}{A variable in the dataframe with which the data set should
be distinct by to avoid a bias towards experiments with many observations.}

\item{data}{An augmented data frame ready for modelling.}

\item{summarise_by}{Which variable(s) a user whishes to get a summary of. To
give a list of variables write \code{rlang::exprs(var1, var2)}}
}
\value{
The function returns a summarised count of whichever variable
is provided in the argument \code{summarise_by} stratified on whichever
variable is provided is provided in the argument \code{distinct}.
}
\description{
Title
}
