% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/410Model_summarise_with_filter.R
\name{summarise_with_filter}
\alias{summarise_with_filter}
\title{Title}
\usage{
summarise_with_filter(.data, ..., identifier = barcode)
}
\arguments{
\item{...}{Which variable(s) a user wishes to get a summary of. To
give a list of variables write them individually.}

\item{identifier}{A variable in the data frame with which the data set should
be distinct by to avoid a bias towards experiments with many observations.
Needs to be specified, even if set to default.}

\item{data}{An augmented data frame ready for modelling.}
}
\value{
The function returns a summarised count of whichever variable
is provided in the argument \code{...} stratified on whichever
variable is provided in the argument \code{distinct}.
}
\description{
Title
}
\seealso{
Other Modelling functions: 
\code{\link{alpha_beta_consistency}()},
\code{\link{alpha_beta_pair_distribution}()},
\code{\link{logo_plot}()},
\code{\link{multiple_relevant_binders}()},
\code{\link{relevant_binder_frequency_plot}()}
}
\concept{Modelling functions}
