% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/270Clean_tidy_allele_names.R
\name{tidy_allele_names}
\alias{tidy_allele_names}
\title{Title}
\usage{
tidy_allele_names(
  .data,
  names_into = c("allele", "peptide", "peptide_source"),
  separator = "_"
)
}
\arguments{
\item{.data}{is a data structure which need preparation}
}
\value{
Return a data frame with the "allele_info" split into three variables:
\itemize{
\item "allele"
\item "peptide"
\item "peptide_source"
While keeping the original variable to use for modelling.
}
}
\description{
Title
}
\seealso{
Other Cleaning functions: 
\code{\link{add_chain_ident_remove_prefix}()},
\code{\link{pivot_longer_TCR_sequences}()},
\code{\link{pivot_longer_allele}()},
\code{\link{remove_unnecessary_columns}()},
\code{\link{run_all_prep}()},
\code{\link{split_TCR_sequences}()}
}
\concept{Cleaning functions}
