% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vdj_v1_hs_aggregated_donor3_binarized_matrix}
\alias{vdj_v1_hs_aggregated_donor3_binarized_matrix}
\title{CD8+ T-cells of Healthy donor 3}
\format{
A data frame with 37824 rows and 118 variables:
\describe{
\item{barcode}{basepair sequence of peptide}
\item{donor}{name of donor}
\item{cell_clono_cdr3_aa}{amino acid sequence of CDR3 alpha and beta chains}
\item{cell_clono_cdr3_nt}{nucleotide sequence of CDR3 alpha and beta chains}
}
}
\source{
\url{https://www.10xgenomics.com/resources/datasets/cd-8-plus-t-cells-of-healthy-donor-3-1-standard-3-0-2}
}
\usage{
vdj_v1_hs_aggregated_donor3_binarized_matrix
}
\description{
A dataset containing roughly 40 000 TCR alpha- and beta sequences for
Donor 3 along with different coding regions.
}
\keyword{datasets}
