% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/450Model_alpha_beta_pair_distributions.R
\name{alpha_beta_pair_distribution}
\alias{alpha_beta_pair_distribution}
\title{Bar plot showing distribution of TCR-sequence promiscuity}
\usage{
alpha_beta_pair_distribution(.data, pMHC_filter = NULL)
}
\arguments{
\item{.data}{An prepared data frame ready for modelling.}

\item{pMHC_filter}{A pMHC to use for filtering, to only include that specific
pMHC from the data set. Default is \code{NULL}, i.e. no filtering. Must be
supplied as string.}
}
\value{
Returns a bar plot with a frequency distribution of the different
combinations of alpha- beta pairs.
}
\description{
\code{alpha_beta_pair_distribution()} takes as input a prepared data frame. It plots
the distribution of categories found in the column \code{TCR_combination}.
A total of four categories exist:
* Only one alpha chain and no beta chain
* No alpha chain and one beta chain
* One alpha- and one beta chain
* Other combinations
}
\examples{
# The function is run by simply piping `.data` into it:
data_donor_one_tidy \%>\%
    alpha_beta_pair_distribution()

# A specific pMHC can be specified to show distribution of only that pMHC:
data_donor_one_tidy \%>\%
    alpha_beta_pair_distribution(pMHC_filter = "A0101_VTEHDTLLY_IE-1_CMV")

}
\seealso{
Other Modelling functions: 
\code{\link{alpha_beta_consistency}()},
\code{\link{logo_plot}()},
\code{\link{multiple_relevant_binders}()},
\code{\link{relevant_binder_frequency_plot}()},
\code{\link{summarise_with_filter}()}
}
\concept{Modelling functions}
