% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_combined_tidy}
\alias{data_combined_tidy}
\title{Prepared data from all four donors}
\format{
A data frame with 2.1 mill rows and 27 columns
}
\source{
Prepared data from all four raw data files:
\itemize{
\item \code{data_donor_one_raw}. raw data file from donor one from \url{https://cf.10xgenomics.com/samples/cell-vdj/3.0.2/vdj_v1_hs_aggregated_donor1/vdj_v1_hs_aggregated_donor1_binarized_matrix.csv}
\item \code{data_donor_two_raw}. raw data file from donor two from \url{https://cf.10xgenomics.com/samples/cell-vdj/3.0.2/vdj_v1_hs_aggregated_donor2/vdj_v1_hs_aggregated_donor2_binarized_matrix.csv}
\item \code{data_donor_three_raw}. raw data file from donor three from \url{https://cf.10xgenomics.com/samples/cell-vdj/3.0.2/vdj_v1_hs_aggregated_donor3/vdj_v1_hs_aggregated_donor3_binarized_matrix.csv}
\item \code{data_donor_four_raw}. raw data file from donor four from \url{https://cf.10xgenomics.com/samples/cell-vdj/3.0.2/vdj_v1_hs_aggregated_donor4/vdj_v1_hs_aggregated_donor4_binarized_matrix.csv}
}
}
\usage{
data_combined_tidy
}
\description{
The data frame contains a prepared version of the data from all donors.
Raw data files were combined and piped through all cleaning and
augmenting functions.
}
\keyword{datasets}
