% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_donor_one}
\alias{data_donor_one}
\title{Prepared data from donor one}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 512264 rows and 27 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
data_donor_one
}
\description{
The data frame contains a prepared version of the data from donor one
}
\keyword{datasets}
