% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/320Augment_evaluate_binder.R
\name{evaluate_binder}
\alias{evaluate_binder}
\title{Calculate whether a binding is significant or not based on thresholds}
\usage{
evaluate_binder(
  .data,
  identifier = barcode,
  UMI_count_min = 10,
  non_specific_UMI_count_min = 5
)
}
\arguments{
\item{.data}{A cleaned data frame ready for augmentation.}

\item{identifier}{Name of the column in the data set which identifies each
experiment. Default is set to \code{barcode}.}

\item{UMI_count_min}{An integer for the minimum number of UMI-counts to be
observed before a binding can be assessed as relevant. The default
is set to \code{10} by 10X Genomics.}

\item{non_specific_UMI_count_min}{An integer which is multiplied with the
maximum value for a non-specific binder. The product is used as
another minimum requirement for the UMI-count to be relevant.
Default is set to \code{5} by 10X Genomics.}
}
\value{
\code{.data} but with a new column which is a \code{boolean}. The \code{boolean} is
\code{TRUE} if the binding between the pMHC and TCR-sequence is relevant.
Otherwise returns \code{FALSE}.
}
\description{
\code{evaluate_binder()} takes a set of thresholds and based on these evaluates
whether a certain UMI-count is to be considered relevant or not. The
UMI-count has to be larger than \code{UMI_count_min}, and multiple times
larger than the largest non-specific binder.
}
\examples{
# The function can be run without any arguments except `.data` if `.data` is
# a tidy data frame such as:
data_donor_one_tidy \%>\%
    evaluate_binder()

# Otherwise, an un-tidy data frame can be piped through the wrapper `?run_all_prep`:
data_donor_one_raw \%>\%
    run_all_prep() \%>\%
    evaluate_binder

# If supplied, the thresholds can be changed from the 10X Genomics standard as:
data_donor_one_tidy \%>\%
    evaluate_binder(UMI_count_min = 20,
                    non_specific_UMI_count_min = 10)

}
\seealso{
Other Preparation functions: 
\code{\link{run_all_prep}()}
}
\concept{Preparation functions}
