% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/320Augment_evaluate_binder.R
\name{evaluate_binder}
\alias{evaluate_binder}
\title{Title}
\usage{
evaluate_binder(
  .data,
  identifier = barcode,
  threshold = 10,
  non_specific_threshold = 5
)
}
\arguments{
\item{.data}{A cleaned data frame ready for augmentation.}

\item{identifier}{The column in the data set which identifies each
experiment. Default is set to \code{barcode}.}

\item{threshold}{An integer for the minimum number of UMI counts to be
observed before a binding can be assessed as relevant. The default
is set to \code{10}.}

\item{non_specific_threshold}{An integer which is multiplied with the
maximum value for a non-specific binder. The product is used as
another minimum requirement for the UMI-count the be relevant.
Default is set to \code{5}.}
}
\value{
The same data frame, but with a new column which is a \code{boolean}.
The \code{boolean} is \code{TRUE} if the binding between the peptide and
TCR-sequence is relevant. Otherwise returns \code{FALSE}.
}
\description{
Title
}
\concept{Augmenting functions}
