% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/420Model_logo_plot.R
\name{logo_plot}
\alias{logo_plot}
\title{Title}
\usage{
logo_plot(
  .data,
  chain_filter = "alpha",
  identifier = barcode,
  sequence_length = 9
)
}
\arguments{
\item{.data}{An prepared data frame ready for modelling.}

\item{chain_filter}{A string value to identify which chain \code{alpha} or
\code{beta} used for the logo plot. Default is set to \code{alpha}.}

\item{identifier}{A variable in the data frame with which the data set should
be distinct by to avoid a bias towards experiments with many observations.}

\item{sequence_length}{An integer for the length of \code{TCR_sequence} desired
for the logo plot.}
}
\value{
A logoplot of the specific chain of TCR-sequences specified by the
argument \code{chain_filter}. To avoid duplicated cells, the input is made
distinct with the argument \code{identifier}, with the default \code{barcode}.
}
\description{
Title
}
\seealso{
Other Modelling functions: 
\code{\link{HLA_match_distribution}()},
\code{\link{alpha_beta_distinctiveness}()},
\code{\link{alpha_beta_pair_distribution}()},
\code{\link{count_binding_pr_allele}()},
\code{\link{multiple_relevant_binders}()},
\code{\link{old_relevant_binder_frequency_plot}()},
\code{\link{relevant_binder_frequency_plot}()},
\code{\link{summarise_with_filter}()}
}
\concept{Modelling functions}
