% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/420Model_logo_plot.R
\name{logo_plot}
\alias{logo_plot}
\title{Title}
\usage{
logo_plot(.data_aug, chain_filter = "alpha", distinct_by = barcode)
}
\arguments{
\item{distinct_by}{}
}
\value{
A logoplot of the specific chain of TCR-sequences specified by the
argument \code{chain_filter}. To avoid duplicated cells, the input is made
distinct with the argument \code{distinct_by}, with the default \code{barcode}.
}
\description{
Title
}
