% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250Clean_pivot_longer_allele.R
\name{pivot_longer_allele}
\alias{pivot_longer_allele}
\title{Title}
\usage{
pivot_longer_allele(data)
}
\arguments{
\item{data}{is a data structure which need cleaning}
}
\value{
A pivot longer version of the data with UMI counts as rows instead
of columns.
}
\description{
Title
}
\seealso{
Other Cleaning functions: 
\code{\link{RunAllClean}()},
\code{\link{ensure_one_alpha_beta_pair}()},
\code{\link{remove_binders}()},
\code{\link{remove_prefix}()},
\code{\link{split_TCR_sequences}()}
}
\concept{Cleaning functions}
