% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/230Clean_pivot_longer_TCR_sequence.R
\name{pivot_longer_TCR_sequences}
\alias{pivot_longer_TCR_sequences}
\title{Title}
\usage{
pivot_longer_TCR_sequences(data)
}
\value{
A pivot longer version of the data with TCR sequences as rows instead
of columns.
}
\description{
Title
}
\seealso{
Other Cleaning functions: 
\code{\link{add_chain_ident_remove_prefix}()},
\code{\link{pivot_longer_allele}()},
<<<<<<< HEAD
\code{\link{remove_unnecessary_binders}()},
\code{\link{run_all_clean}()},
\code{\link{split_TCR_sequences}()}
=======
\code{\link{remove_unnecessary_columns}()},
\code{\link{run_all_clean}()},
\code{\link{split_TCR_sequences}()},
\code{\link{tidy_allele_names}()}
>>>>>>> 8e8c7423f3b50a390b49a73d2d56a44ddf2fa606
}
\concept{Cleaning functions}
