% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/230Clean_pivot_longer_TCR_sequence.R
\name{pivot_longer_TCR_sequences}
\alias{pivot_longer_TCR_sequences}
\title{Title}
\usage{
pivot_longer_TCR_sequences(data)
}
\arguments{
\item{data}{is a data structure which need cleaning}
}
\value{
A pivot longer version of the data with TCR sequences as rows instead
of columns.
}
\description{
Title
}
\seealso{
Other Cleaning functions: 
\code{\link{add_chain_ident_remove_prefix}()},
\code{\link{pivot_longer_allele}()},
\code{\link{remove_unnecessary_columns}()},
\code{\link{run_all_clean}()},
\code{\link{split_TCR_sequences}()},
\code{\link{tidy_allele_names}()}
}
\concept{Cleaning functions}
