% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/430Model_relevant_binder_frequency_plot.R
\name{relevant_binder_frequency_plot}
\alias{relevant_binder_frequency_plot}
\title{Title}
\usage{
relevant_binder_frequency_plot(.data, identifier = barcode)
}
\arguments{
\item{.data}{An augmented data frame ready for modelling.}

\item{identifier}{A variable in the data frame with which the data set should
be distinct by to avoid a bias towards experiments with many observations.
Needs to be specified, even if set to default.}
}
\value{
An interactive plotly plot with different relevant alpha:beta pairs,
and pMHC. The size of dots depends on the frequency of that specific
interaction occuring compared to other pMHC which bind the same
alpha:beta pair.
}
\description{
Title
}
\seealso{
Other Modelling functions: 
\code{\link{alpha_beta_pair_distribution}()},
\code{\link{logo_plot}()},
\code{\link{multiple_specific_binders}()},
\code{\link{summarise_with_filter}()}
}
\concept{Modelling functions}
