% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/210Clean_remove_unnecessary_columns.R
\name{remove_unnecessary_columns}
\alias{remove_unnecessary_columns}
\title{Title}
\usage{
remove_unnecessary_columns(
  .data,
  col1 = "_binder",
  col2 = "cell_clono_cdr3_nt"
)
}
\arguments{
\item{.data}{is a data structure which need preparation}
}
\value{
Returns the input data, but without all the binding assessments and
nucleotide sequences
}
\description{
Title
}
\seealso{
Other Cleaning functions: 
\code{\link{add_chain_ident_remove_prefix}()},
\code{\link{pivot_longer_TCR_sequences}()},
\code{\link{pivot_longer_allele}()},
\code{\link{run_all_prep}()},
\code{\link{split_TCR_sequences}()},
\code{\link{tidy_allele_names}()}
}
\concept{Cleaning functions}
