% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/200Clean_run_all_prep.R
\name{run_all_prep}
\alias{run_all_prep}
\title{Run all preparation functions}
\usage{
run_all_prep(.data)
}
\arguments{
\item{.data}{A data frame which need preparation before modelling. Four such
data frames are already present in this package:
* \code{data_donor_one_raw}
* \code{data_donor_two_raw}
* \code{data_donor_three_raw}
* \code{data_donor_four_raw}}
}
\value{
An object which is now prepared data frame, of the same type as \code{.data}.
The output has the following differences:
\itemize{
\item Unnecessary columns are removed
\item TCR-sequences and pMHC are now tidy, and pivot longer
\item Several columns are added to aid modelling
}
}
\description{
\code{run_all_prep()} prepares the data frame it is given which is done through
a series of functions. Firstly, the columns which aren't needed are removed.
Then, the TCR-sequences are made tidy by splitting them and pivot longer.
To keep track of their type - alpha or beta - a column is added specifying
exactly that. The pMHC are made tidy and pivot longer as for the TCR-sequences.
Further, the maximum UMI-count of all non-specific binders is added to
enable an evaluation of the relevancy of a TCR-sequence immediately after.
Lastly, a categorical identifier is added telling about the promiscuity
of a TCR-sequence.
}
\details{
The following functions are run:
\itemize{
\item \code{remove_unnecessary_columns()}
\item \code{find_non_promiscuous_split_TCR_sequences()}
\item \code{pivot_longer_TCR_sequences()}
\item \code{add_chain_ident_remove_prefix()}
\item \code{pivot_longer_pMHC()}
\item \code{tidy_allele_names()}
\item \code{add_max_non_specific_binder()}
\item \code{evaluate_binder()}
\item \code{add_TCR_combination_identifier()}
}
}
\examples{
# Simply supply the function with the name of a data frame:
run_all_prep(data_donor_one_raw)

# Or by using the pipe from `magrittr`:
data_donor_one_raw \%>\%
    run_all_prep()

}
\seealso{
Other Preparation functions: 
\code{\link{evaluate_binder}()}
}
\concept{Preparation functions}
