% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/200Clean_RunAll.R
\name{RunAllClean}
\alias{RunAllClean}
\title{Run all other clean-functions}
\usage{
RunAllClean(data)
}
\arguments{
\item{data}{is a data structure which need cleaning}
}
\value{
A cleaned data structure, of the same type as the input.
The output has the following differences:
\itemize{
\item Observations are removed if they do not have a TRA:TRB pair
\item Two new columns are added, one for TRA and one for TRB
\item The binding assessments are removed
}
}
\description{
\code{RunAllClean()} cleans the data structure it is given by removing
incomplete/wrong TRA:TRB sequence pairs and removing all the binding assessments.
The following functions are run:
\itemize{
\item 210Clean_EnsureOneAlphaBetaPair.R
\item 220Clean_SplitTCRSequences.R
\item 230Clean_RemovePrefix.R
\item 240Clean_RemoveBinders.R
}
}
\examples{
RunAllClean(vdj_v1_hs_aggregated_donor1_binarized_matrix.rda)
}
\seealso{
Other Cleaning functions: 
\code{\link{EnsureOneAlphaBetaPair}()}
}
\concept{Cleaning functions}
