% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/220Clean_split_TCR_sequences.R
\name{split_TCR_sequences}
\alias{split_TCR_sequences}
\title{Title}
\usage{
split_TCR_sequences(.data, TCR_col = cell_clono_cdr3_aa, separator = ";")
}
\arguments{
\item{.data}{is a data structure which need cleaning}
}
\value{
Returns the input data, but "cell_clono_cdr3_aa" is split for each
TCR-sequence.
}
\description{
Title
}
\seealso{
Other Cleaning functions: 
\code{\link{add_chain_ident_remove_prefix}()},
\code{\link{pivot_longer_TCR_sequences}()},
\code{\link{pivot_longer_allele}()},
\code{\link{remove_unnecessary_columns}()},
\code{\link{run_all_clean}()},
\code{\link{tidy_allele_names}()}
}
\concept{Cleaning functions}
