% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/220Clean_split_TCR_sequences.R
\name{split_TCR_sequences}
\alias{split_TCR_sequences}
\title{Title}
\usage{
split_TCR_sequences(data)
}
\arguments{
\item{data}{is a data structure which need cleaning}
}
\value{
The input data, but each TCR-pair is split into a TRA and TRB column
}
\description{
Title
}
\seealso{
Other Cleaning functions: 
\code{\link{RunAllClean}()},
\code{\link{ensure_one_alpha_beta_pair}()},
\code{\link{pivot_longer_allele}()},
\code{\link{remove_binders}()},
\code{\link{remove_prefix}()}
}
\concept{Cleaning functions}
