% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/410Model_summarise_binders.R
\name{summarise_binders}
\alias{summarise_binders}
\title{Title}
\usage{
summarise_binders(
  .data_aug,
  distinct_by = barcode,
  summarise_by = rlang::exprs(allele, peptide, peptide_source)
)
}
\arguments{
\item{.data_aug}{}

\item{distinct_by}{}

\item{summarise_by}{}
}
\value{
The function returns a summarised count of whichever variable
is provided in the argument \code{summarise_by} stratified on whichever
variable is provided is provided in the argument \code{distinct}.
}
\description{
Title
}
