% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/410Model_summarise_with_filter.R
\name{summarise_with_filter}
\alias{summarise_with_filter}
\title{Count of relevant binders stratified on user input}
\usage{
summarise_with_filter(
  .data,
  summarise_by = c("allele", "peptide", "peptide_source"),
  identifier = barcode
)
}
\arguments{
\item{.data}{An prepared data frame ready for modelling.}

\item{summarise_by}{The column(s) a user wishes to get a summary of. Provide
as a string or vector of strings. Default is the columns prepared from
pMHC.}

\item{identifier}{A variable in the data frame with which the data set should
be distinct by to avoid a bias towards experiments with many observations.}
}
\value{
A table with a count of relevant binders stratified on \code{summarise_by}.
}
\description{
\code{summarise_with_filter()} takes a prepared data frame as input, and count
the number of pMHC which are evaluated as relevant by \code{?evaluate_binder}.
The counting is grouped by columns provided by the user through the argument
\code{...}.
}
\examples{
# A prepared data frame needs to be provided. Can either be done by using
# the already cleaned data frames or the raw data files piped through `?run_all_prep`
data_donor_one_tidy \%>\%
    summarise_with_filter()

data_donor_one_raw \%>\%
    run_all_prep() \%>\%
    summarise_with_filter()

# The stratification can be done for one or several columns:
data_donor_one_tidy \%>\%
    summarise_with_filter(summarise_by = "allele")

data_donor_one_tidy \%>\%
    summarise_with_filter(summarise_by = c("allele",
                                           "peptide_source"))

}
\seealso{
Other Modelling functions: 
\code{\link{alpha_beta_consistency}()},
\code{\link{alpha_beta_pair_distribution}()},
\code{\link{count_binding_pr_allele}()},
\code{\link{logo_plot}()},
\code{\link{multiple_relevant_binders}()},
\code{\link{predict_allele_typing}()},
\code{\link{predict_virus_status}()},
\code{\link{relevant_binder_frequency_plot}()}
}
\concept{Modelling functions}
