% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/270Clean_tidy_allele_names.R
\name{tidy_allele_names}
\alias{tidy_allele_names}
\title{Title}
\usage{
tidy_allele_names(data)
}
\value{
Return a data frame with the "allele_info" split into three columns:
\itemize{
\item "allele"
\item "peptide"
\item "peptide_source"
}
}
\description{
Title
}
\seealso{
Other Cleaning functions: 
\code{\link{add_chain_ident_remove_prefix}()},
\code{\link{pivot_longer_TCR_sequences}()},
\code{\link{pivot_longer_allele}()},
<<<<<<< HEAD:man/allele_names.Rd
\code{\link{remove_unnecessary_binders}()},
=======
\code{\link{remove_unnecessary_columns}()},
>>>>>>> 8e8c7423f3b50a390b49a73d2d56a44ddf2fa606:man/tidy_allele_names.Rd
\code{\link{run_all_clean}()},
\code{\link{split_TCR_sequences}()}
}
\concept{Cleaning functions}
