% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/480Model_umap_of_protein_expressions.R
\name{umap_of_protein_expressions}
\alias{umap_of_protein_expressions}
\title{Umap of the protein expression levels}
\usage{
umap_of_protein_expressions(.data, identifier = barcode, color_by = "CD45RA")
}
\arguments{
\item{.data}{An prepared data frame ready for modelling.}

\item{identifier}{A variable in the data frame with which the data set should
be distinct by to avoid a bias towards experiments with many observations.}

\item{color_by}{A variable name to choose which protein expression level to
color by.}
}
\value{
A scatter plot to show tendencies of the gene expression levels.
}
\description{
\code{umap_of_protein_expressions()} takes a prepared data frame, \code{.data}, and
projects the data points via the already made models included in the
package.
}
\examples{
# A prepared data frame is simply piped through the function:
data_combined_tidy \%>\%
    umap_of_protein_expressions()

# The protein to color by can be changed:
data_combined_tidy \%>\%
    umap_of_protein_expressions(color_by = "CD3")

}
\seealso{
Other Modelling functions: 
\code{\link{alpha_beta_distinctiveness}()},
\code{\link{alpha_beta_pair_distribution}()},
\code{\link{boxplot_protein_expressions}()},
\code{\link{count_binding_pr_allele}()},
\code{\link{percentage_rows_kept}()},
\code{\link{relevant_binder_frequency_plot}()},
\code{\link{summarise_with_filter}()}
}
\concept{Modelling functions}
